ajout exe6 > liste
