+++ src/pages/Exercice6.vue
@@ -1,24 +1,29 @@
 <template>
   <v-container max-width="700">
-    <h1>Exercice 7</h1>
-    <div class="exe-objectifs">
-      <h2>Objectifs</h2>
-      <ul>
-        <li>Découvrir et utiliser les hooks du <strong>cycle de vie</strong> dans Vue.js.</li>
-        <li>Utiliser <code>onMounted</code> pour exécuter une action lorsque le composant est monté.</li>
-        <li>Utiliser <code>onUpdated</code> pour détecter une mise à jour globale du composant.</li>
-        <li>Utiliser <code>onUnmounted</code> pour nettoyer des actions, comme annuler un intervalle.</li>
-      </ul>
-    </div>
-    <v-divider class="my-4" />
+    <!-- Donnée de l'exercice -->
+    <exercice5-donnee />
+    <!-- Zone de travail pour l'exercice -->
     <div class="exe-zone">
       <h2>Zone d'exercice</h2>
       <v-card class="mx-auto my-6 pa-2" max-width="500">
-        <v-card-title>Cycle de vie de Vue.js</v-card-title>
+        <v-card-title>Saisie surveillée</v-card-title>
+
         <v-card-text>
-          <p>Compteur : <strong>{{ counter }}</strong></p>
-          <p>Dernière mise à jour : {{ updatedTimestamp }}</p>
-          <v-btn color="primary" @click="incrementCounter">Incrémenter</v-btn>
+          <v-alert type="success" class="mb-2">
+            Vous avez mentionné "Pokémon" !
+          </v-alert>
+
+          <v-text-field
+            v-model="userInput"
+            label="Tapez quelque chose"
+            placeholder="Essayez d'écrire Pokémon"
+            outlined
+          />
+
+          <v-card-subtitle>
+            Nombre de caractères :
+            *** CARACTÈRES SAISIS ***
+          </v-card-subtitle>
         </v-card-text>
       </v-card>
     </div>
@@ -26,50 +31,17 @@
 </template>
 
 <script setup>
-import { ref, onMounted, onUpdated, onUnmounted } from 'vue';
+// Importation du composant contenant la donnée de l'exerciced
+import Exercice5Donnee from "@/components/donnees/Exercice5Donnee.vue";
+// Importation de la fonction réactive ref
+import { ref } from 'vue';
 
-// Variables réactives
-const counter = ref(0);
-const updatedTimestamp = ref('');
-let intervalId = null;
+// Variable réactive pour la saisie utilisateur
+const userInput = ref('');
 
-// Hook exécuté lorsque le composant est monté
-onMounted(() => {
-  console.log('Composant monté !');
-  // Simule un intervalle qui met à jour le compteur automatiquement
-  intervalId = setInterval(() => {
-    counter.value++;
-  }, 2000);
-});
+// Longueur maximale autorisée
+const MAX_LENGTH = 20;
 
-// Hook exécuté lorsque le composant est mis à jour (DOM mis à jour)
-onUpdated(() => {
-  console.log('Le composant a été mis à jour !');
-  const now = new Date();
-  updatedTimestamp.value = `${now.toLocaleTimeString()}`;
-});
-
-// Hook exécuté lorsque le composant est démonté
-onUnmounted(() => {
-  console.log('Composant démonté !');
-  // Nettoie l'intervalle pour éviter une fuite de mémoire
-  clearInterval(intervalId);
-});
-
-// Fonction pour incrémenter manuellement le compteur
-const incrementCounter = () => {
-  counter.value++;
-};
+// Variable réactive pour indiquer si "Pokémon" est présent
+const containsPokemon = ref(false);
 </script>
-
-<style scoped>
-.exe-objectifs {
-  background-color: #f5f5f5;
-  padding: 16px;
-  border-radius: 8px;
-}
-
-.exe-zone {
-  margin-top: 16px;
-}
-</style>
